+++ src/utils/copy/copy_vector.cpp
@@ -43,7 +43,8 @@ template <typename T> void vector<T>::operator=(const vector<T> &vec) {
 template void vector<double>::operator=(const vector<double> &vec);
 template void vector<float>::operator=(const vector<float> &vec);
 
-template <typename T> void vector<T>::operator=(const view1D<vector<T>,T> &vec) {
+template <typename T>
+void vector<T>::operator=(const view1D<vector<T>, T> &vec) {
   Logger &logger = Logger::get_instance();
   logger.util_in(monolish_func);
 
@@ -54,19 +55,23 @@ template <typename T> void vector<T>::operator=(const view1D<vector<T>,T> &vec) 
   // gpu copy and recv
   if (vec.get_device_mem_stat()) {
 #if MONOLISH_USE_GPU
-    internal::vcopy(vec.size(), vec.data()+vec.get_offset(), val.data(), true);
+    internal::vcopy(vec.size(), vec.data() + vec.get_offset(), val.data(),
+                    true);
 #endif
   } else {
-    internal::vcopy(vec.size(), vec.data()+vec.get_offset(), val.data(), false);
+    internal::vcopy(vec.size(), vec.data() + vec.get_offset(), val.data(),
+                    false);
   }
 
   logger.util_out();
 }
 
-template void vector<double>::operator=(const view1D<vector<double>,double> &vec);
-template void vector<float>::operator=(const view1D<vector<float>,float> &vec);
+template void
+vector<double>::operator=(const view1D<vector<double>, double> &vec);
+template void vector<float>::operator=(const view1D<vector<float>, float> &vec);
 
-template <typename T> void vector<T>::operator=(const view1D<matrix::Dense<T>,T> &vec) {
+template <typename T>
+void vector<T>::operator=(const view1D<matrix::Dense<T>, T> &vec) {
   Logger &logger = Logger::get_instance();
   logger.util_in(monolish_func);
 
@@ -77,15 +82,19 @@ template <typename T> void vector<T>::operator=(const view1D<matrix::Dense<T>,T>
   // gpu copy and recv
   if (vec.get_device_mem_stat()) {
 #if MONOLISH_USE_GPU
-    internal::vcopy(vec.size(), vec.data()+vec.get_offset(), val.data(), true);
+    internal::vcopy(vec.size(), vec.data() + vec.get_offset(), val.data(),
+                    true);
 #endif
   } else {
-    internal::vcopy(vec.size(), vec.data()+vec.get_offset(), val.data(), false);
+    internal::vcopy(vec.size(), vec.data() + vec.get_offset(), val.data(),
+                    false);
   }
 
   logger.util_out();
 }
 
-template void vector<double>::operator=(const view1D<matrix::Dense<double>,double> &vec);
-template void vector<float>::operator=(const view1D<matrix::Dense<float>,float> &vec);
+template void
+vector<double>::operator=(const view1D<matrix::Dense<double>, double> &vec);
+template void
+vector<float>::operator=(const view1D<matrix::Dense<float>, float> &vec);
 } // namespace monolish